backend/requirements.txt
Django
django-cors-headers
mysqlclient
cryptography
mysql-connector-python
djangorestframework
djangorestframework